merge

+++ moteur 2D/moteur 2D/game.h
@@ -14,7 +14,7 @@ class Game {
 	float elder_scrolls; //Euh actually c est une ref a la licence elder scrolls mais en vrai la variable c est pour le debut du scrolling.
 public:
 	void Collisions(float now, SDL_Renderer* renderer, std::vector<Block*>& blocks, Character& ch);
-	void GameRenderer(bool gameStart, float winWidth, SDL_Renderer* renderer, Character& ch,
+	void GameRenderer(bool gameStart, int winWidth, SDL_Renderer* renderer, Character& ch,
 		std::vector<Block*>& blocks, Background& bg, Menu& menu, Button* exit, Button* start, Button* gameOver, 
 		std::vector<SDL_Texture*>& gridTextures);
 	void Update(float dt, float gameTime, bool& gameStart, Character& ch,
